refactor: repository 삭제
