+++ src/main/java/racingcar/domain/CarName.java
@@ -34,8 +34,12 @@ public class CarName {
 
     @Override
     public boolean equals(Object diffCarName) {
-        if (this == diffCarName) return true;
-        if (diffCarName == null || getClass() != diffCarName.getClass()) return false;
+        if (this == diffCarName) {
+            return true;
+        }
+        if (diffCarName == null || getClass() != diffCarName.getClass()) {
+            return false;
+        }
         CarName carName = (CarName) diffCarName;
         return Objects.equals(name, carName.name);
     }
